Simplify input arguments for the two-index and F-constraints
Both of these require the spatial derivative of the gauge
source function H. As the spacetime derivative of H is almost
always added to the evolution databox anyway, it is simpler
for these constraints' computing methods and compute tags
to directly accept the spacetime derivative of H as input
and extract the spatial derivative internally. This commit
makes this simplification.