Change AlgorithmImpl's GlobalCache pointer to a proxy
Previously, AlgorithmImpl stored a GlobalCache pointer to enable the
DataBox to access GlobalCache items.

When deserializing an AlgorithmImpl component (e.g. after a load
balance or while restoring from a checkpoint file), the GlobalCache
pointer needed to be updated to point to the local GlobalCache on
the new CPU. Obtaining this local GlobalCache pointer proved to be
troublesome in a restart from checkpoint, because (depending on
details of the Charm++ runtime) the GlobalCache proxy may not yet be
in a valid state for calling .ckLocalBranch() until all components
are restored.

This commit changes AlgorithmImpl to store a GlobalCache proxy
instead of a pointer. This enables the DataBox to compute the
GlobalCache pointer only when it is needed, which delays the call
until after deserialization is complete. This change should have no
effect on LB migrations, but enables deserializing from checkpoints.