Terminate CCE when waiting for worldtube data
The `receive_data` calls appropriately restart the algorithm execution.

This change allows phase changes from quiescence detection due to
inactivity (i.e. no more data being supplied by the GH elements); without
terminating at the stage in which the component waits for data, the Algorithm
generates an error if a phase change occurs while the CCE component is inactive.

We want to permit such phase changes, because the CCE component is in a safe
state for balancing or checkpointing when it is inactive and waiting for data.